chore: restructure autora.state, replacing autora.workflow.state
