+++ src/autora/workflow/state/__init__.py
@@ -1,4 +1,13 @@
-from .delta import Delta, State, extend, wrap_to_use_state
-from .history import History
-from .param import resolve_state_params
-from .snapshot import Snapshot
+import logging
+
+from autora.state.delta import Delta, State, extend, wrap_to_use_state
+from autora.state.history import History
+from autora.state.param import resolve_state_params
+from autora.state.snapshot import Snapshot
+
+_logger = logging.getLogger(__name__)
+
+_logger.warning(
+    "`autora.workflow.state` will be removed in a future release. "
+    "Use `autora.state` in future."
+)
